update hook code
