+++ .gitignore
@@ -1,4 +1,5 @@
+.vscode/
+.cache/
 build/
 out/
-.vscode/
 compile_commands.json
