+++ CMakeLists.txt
@@ -21,7 +21,8 @@ if (CMAKE_CXX_COMPILER_ID MATCHES "Clang")
 endif()
 
 add_subdirectory(aosp_file)
-add_subdirectory(backtrace_helper)
+add_subdirectory(unwindstack/cmake)
+add_subdirectory(backtrace)
 
 if (ANDROID_ABI MATCHES "arm64-v8a")
     set(VNDK_LIB_PATH "${CMAKE_CURRENT_SOURCE_DIR}/aosp_file/com.android.vndk.v30/lib64")
@@ -32,9 +33,9 @@ else ()
 endif ()
 
 # 添加 .so
-add_library(alloc_hook SHARED ${CMAKE_SOURCE_DIR}/src/host_alloc_hook.cpp ${CMAKE_SOURCE_DIR}/src/dma_alloc_hook.cpp)
+add_library(alloc_hook SHARED ${CMAKE_SOURCE_DIR}/src/alloc_hook.cpp)
 # 链接需要的库和头文件
-target_link_libraries(alloc_hook PRIVATE backtrace_helper)
+target_link_libraries(alloc_hook PRIVATE helper)
 # 安装到 out/lib 目录
 install(TARGETS alloc_hook DESTINATION ${CMAKE_INSTALL_PREFIX}/out/lib)
